requirements.txt
alembic==1.4.3
fastapi==0.62.0
FastAPI-SQLAlchemy==0.2.1
pydantic==1.7.3
python-dotenv==0.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
SQLAlchemy==1.3.20
uvicorn==0.12.3
